move convert props to session
